+++ docs/whatsnew/changelog.rst
@@ -26,10 +26,8 @@ New features
 * Filters has been refactored. A new `Filter` processor class allows to define various
   filters and apply them to a dataset. See `Filtering and Smoothing` tutorials and `Filter`
   for details. Note: Backward compatibility is ensured with the previous `smooth` and `savgol_filter` methods.
-
 * A `whittaker` filter has been added to the `Filter` processor class. See `Filtering and Smoothing`
   tutorials and `Filter` for details.
-
 * A new `Baseline` processor class has been added. It allows to perform baseline correction
   on a dataset with multiple algorithms. See :ref:`Baseline` for details.
 
@@ -39,6 +37,7 @@ Bug fixes
 ~~~~~~~~~
 .. Add here new bug fixes (do not delete this comment)
 
+* #687 fixed.
 
 .. section
 
@@ -58,3 +57,5 @@ Breaking changes
 Deprecations
 ~~~~~~~~~~~~
 .. Add here new deprecations (do not delete this comment)
+
+* `parameters` method of Analysis configurables is now deprecated in favor of `params`.